refactor(app): decompose OneStepApp god-object into facade + runtime controllers (#146)
Split src/onestep/app.py (1136->389 lines) into a thin facade delegating to
three stdlib-only runtime controllers (zero new deps):
- runtime/event_hub.py EventHub (108 lines): startup/shutdown hooks, event handlers, emit_event
- runtime/task_ops.py TaskOperations (324 lines): dead-letter replay/discard, one-shot run, capability probes (_task_supports_*)
- runtime/lifecycle.py LifecycleController (649 lines): asyncio state, runner registry, startup/shutdown/serve, drain/pause control-plane snapshots, signal handlers, module-level _open_resource/_close_resource helpers

Public method names/signatures/return structures byte-for-byte unchanged.
Facade keeps read-only _runners/_runner_tasks/_resources properties and
_install_signal_handlers() passthrough because tests/contract reads them directly.
Added AGENTS.md module-boundary note and design spec doc.

Closes #146
